+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [3.3.0] - 2018-10-08
+
+### Added
+
+- Generate All
+- Option to put the method's opening brace on a new line instead of the same line "java.code.generators.methodOpeningBraceOnNewLine"
+
+### Fixed
+
+- selecting any variable marked as protected result in 'undefined'
+
 ## [3.2.2] - 2018-09-04
 
 ### Added